chore: solves comments
