+++ packages/browser/test/unit/tracekit/next.test.ts
@@ -1,14 +1,10 @@
+import { nextStackParser } from '@sentry/nextjs';
 import { createStackParser } from '@sentry/utils';
 import { exceptionFromError } from '../../../src/eventbuilder';
-import {
-  chromeStackLineParser,
-  geckoStackLineParser,
-  nextRoutesStackParser,
-  winjsStackLineParser,
-} from '../../../src/stack-parsers';
+import { chromeStackLineParser, geckoStackLineParser, winjsStackLineParser } from '../../../src/stack-parsers';
 
 const parser = createStackParser(
-  nextRoutesStackParser,
+  nextStackParser,
   chromeStackLineParser,
   geckoStackLineParser,
   winjsStackLineParser,
@@ -28,22 +24,21 @@ describe('Tracekit - Next.js Tests', () => {
 
     const ex = exceptionFromError(parser, NEXT);
 
-    expect(ex).toMatchInlineSnapshot(`
-Object {
-  "stacktrace": Object {
-    "frames": Array [
-      Object {
-        "colno": 126,
-        "filename": "http://localhost:3001/_next/static/chunks/app/%5Blocale%5D/sentery-example-page/%28default%29/sub/page-3d428c1ba734e10f.js",
-        "function": "?",
-        "in_app": true,
-        "lineno": 1,
+    expect(ex).toEqual({
+      value: "Unable to get property 'undef' of undefined or null reference",
+      type: 'foo',
+      stacktrace: {
+        frames: [
+          {
+            filename:
+              'http://localhost:3001/_next/static/chunks/app/%5Blocale%5D/sentery-example-page/%28default%29/sub/page-3d428c1ba734e10f.js',
+            function: '?',
+            lineno: 1,
+            colno: 126,
+            in_app: true,
+          },
+        ],
       },
-    ],
-  },
-  "type": "foo",
-  "value": "Unable to get property 'undef' of undefined or null reference",
-}
-`);
+    });
   });
 });
