+++ packages/browser/test/unit/tracekit/opera.test.ts
@@ -1,13 +1,9 @@
+import { nextStackParser } from '@sentry/nextjs';
 import { createStackParser } from '@sentry/utils';
 import { exceptionFromError } from '../../../src/eventbuilder';
-import {
-  defaultStackParser,
-  nextRoutesStackParser,
-  opera10StackLineParser,
-  opera11StackLineParser,
-} from '../../../src/stack-parsers';
+import { defaultStackParser, opera10StackLineParser, opera11StackLineParser } from '../../../src/stack-parsers';
 
-const operaParser = createStackParser(nextRoutesStackParser, opera10StackLineParser, opera11StackLineParser);
+const operaParser = createStackParser(nextStackParser, opera10StackLineParser, opera11StackLineParser);
 const chromiumParser = defaultStackParser;
 
 describe('Tracekit - Opera Tests', () => {
